fix: eslint error

+++ app/api/reservations/[reservationId]/route.ts
@@ -2,18 +2,21 @@ import { NextResponse } from "next/server";
 import getCurrentUser from "@/app/actions/GetCurrentUser";
 import prisma from "@/app/libs/prismadb";
 
-export async function DELETE(
-  request: Request,
-  { params }: { params: { reservationId: string } }
-) {
+export async function DELETE(request: Request) {
   const currentUser = await getCurrentUser();
 
   if (!currentUser) {
     return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
   }
 
-  const { reservationId } = params;
-  
+  // Extrair `reservationId` da URL
+  const url = new URL(request.url);
+  const reservationId = url.pathname.split('/').pop(); // ou use regex para maior segurança
+
+  if (!reservationId) {
+    return NextResponse.json({ error: "Invalid reservation ID" }, { status: 400 });
+  }
+
   try {
     const reservation = await prisma.reservation.deleteMany({
       where: {
